backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.10.6
python-dotenv==1.0.1

passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.4.0

beanie==1.26.0
motor==3.6.0
